Move metrics into separate package
