+++ qurator/dinglehopper/__init__.py
@@ -1,5 +1,3 @@
 from .ocr_files import *
 from .extracted_text import *
-from .character_error_rate import *
-from .word_error_rate import *
 from .align import *
