+++ qurator/dinglehopper/metrics/__init__.py
@@ -1,0 +1,2 @@
+from .character_error_rate import *
+from .word_error_rate import *
