+++ qurator/dinglehopper/metrics/character_error_rate.py
@@ -6,8 +6,8 @@ from typing import Tuple
 from multimethod import multimethod
 from uniseg.graphemecluster import grapheme_clusters
 
-from .edit_distance import distance
-from .extracted_text import ExtractedText
+from ..edit_distance import distance
+from ..extracted_text import ExtractedText
 
 
 @multimethod
