+++ qurator/dinglehopper/tests/metrics/test_integ_character_error_rate_ocr.py
@@ -6,9 +6,10 @@ import pytest
 from lxml import etree as ET
 from uniseg.graphemecluster import grapheme_clusters
 
-from .. import character_error_rate, page_text, alto_text
+from ... import page_text, alto_text
+from ...metrics import character_error_rate
 
-data_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "data")
+data_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), "../", "data")
 
 
 @pytest.mark.integration
